+++ internal/models/models.go
@@ -3,22 +3,22 @@ package models
 import "time"
 
 type File struct {
-	ID              int
-	JobID           int
+	ID              int64
+	JobID           int64
 	Name            string
 	CreatedAt       time.Time
 	LatestOperation *Operation
 }
 
 type Job struct {
-	ID        int
+	ID        int64
 	CreatedAt time.Time
 }
 
 type Operation struct {
-	ID        int
-	JobID     int
-	FileID    int
+	ID        int64
+	JobID     int64
+	FileID    int64
 	Format    string
 	Quality   int
 	FileName  string
@@ -28,8 +28,8 @@ type Operation struct {
 }
 
 type PlannedFlush struct {
-	ID        int
-	JobID     int
+	ID        int64
+	JobID     int64
 	FlushAt   time.Time
 	CreatedAt time.Time
 }
